% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize the data
Takes in  a vector (atomic or list) or an expression object and return a vector of same length}
\usage{
normalize(x)
}
\arguments{
\item{X:}{a vector (atomic or list) or an expression object}
}
\value{
A normalized vector of same length.
}
\description{
Normalize the data
Takes in  a vector (atomic or list) or an expression object and return a vector of same length
}

