% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cluster.R
\name{show_cluster}
\alias{show_cluster}
\title{Show the clusters of Iris dataset
Takes in the Iris dataframe and show the PCA result.}
\usage{
show_cluster(df)
}
\arguments{
\item{df}{The dataset}
}
\value{
A plot of clusters
}
\description{
Show the clusters of Iris dataset
Takes in the Iris dataframe and show the PCA result.
}

